+++ CHANGELOG.md
@@ -12,8 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - LangevinSolver has been added to the package. This solver is a stochastic solver based on the Langevin equation.
-- Introduced `solve_type` control keyword when calling the solver objects. If it is deliberately set to 'Adam' then the ADAM algorithm method will be executed with relevant hyperparameters. Otherwise, its default value (i.e. solve_type=None) refers to the call for the original solve method. The Adam method consist of three hyperparameters in which beta1 and beta2 are exponential decay rates for the moment estimates and
-alpha is the step size. Please refer to [the paper](https://doi.org/10.48550/arXiv.1412.6980) for more information about Adam algorithm and its hyperparameters.
+- Added a `__call__` method to the solvers. This allows users to call the solver object directly, rather than having to call the `solve` method.
+  - Introduced the `algorithm_parameters` control keyword when calling the solver objects. This allows users to specify a specialized algorithm for the solver to use and provide the parameters of this algorithm to the solver.
+  - Algorithm parameter classes have been added to `ccvm_simulators.solvers.algorithms` to allow users to specify the `algorithm_parameters`.
 
 ### Changed
 - Updated internal data handling that have improved performance and reduced memory usage.